Move classes related to the stringifiers to the `org.spine3.convert` package.
