+++ client/src/main/java/org/spine3/base/Identifiers.java
@@ -27,6 +27,8 @@ import com.google.protobuf.MessageOrBuilder;
 import com.google.protobuf.Timestamp;
 import com.google.protobuf.util.Timestamps;
 import org.spine3.Internal;
+import org.spine3.convert.Stringifier;
+import org.spine3.convert.StringifierRegistry;
 
 import javax.annotation.Nullable;
 import java.util.Collection;
